GetPlugin v.1.0.2 — get via config
GetPlugin being used in the ViewTemplate can now auto-detect what to get
depending on how a tiddler is tagged, applying a defined template with
the ability to get different references for a tiddler being the tag or
having it.

This is extensively used at http://customize.tiddlyspace.com

Also removed some remnant console.log statements.