DiscussPlugin v.0.9.0 first commit
A rather simple plugin to render Disqus in an iframe in a TiddlyWiki on
TiddlySpace.

Will *not* work in a classic TiddlyWiki unless you hook-up a proxy
server via php et. al. and so some url rewriting that captures the
desired individual url's as needed for disqus to work.

Hashes won't do!