DiscussPlugin v.0.9.1 renamed and fix for non-existing tiddlers
Prevents from showing when tiddler doesn't exist yet.

A rather simple plugin to render Disqus in an iframe in a TiddlyWiki on
TiddlySpace.

Will *not* work in a classic TiddlyWiki unless you hook-up a proxy
server via php et. al. and so some url rewriting that captures the
desired individual url's as needed for disqus to work.

Hashes won't do!