TagFiltrPlugin v.1.4.0 — first commit to plugin repo
TagFiltrPlugin is a fork, rewrite extension of Tagsplorer by FND which
on top of its know behavior allows you to predefine prefixes for which
you can easily select and filter by matching tags.

Later it will also be possible to define any tag as a tag category, e.g.
to define...

[[Task]]

...as the category whereas any...

[[#open]]
[[#next]]
[[#waiting]]
[[#done]]

...tagging to it would be listed in the popup. In other words, rather
than define a prefix, such as #, it will be possible to have any tagging
tiddler listed.